#19 Moved the converter to the config package.
