+++ src/main/kotlin/ro/hibyte/betting/config/WebSecurityConfig.kt
@@ -7,7 +7,6 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.web.SecurityFilterChain
-import ro.hibyte.betting.service.JwtAuthConverter
 
 @Configuration
 @EnableWebSecurity
